Jira-ID: CRNT-27199 merged CRNT-27078 from 5.3.19 - SP22 to trunk.
git-svn-id: http://emeafrazerg/svn/ipp/product/trunk/stardust/engine@61237 8100b5e0-4d52-466c-ae9c-bdeccbdeaf6b

+++ stardust-engine-core/src/main/java/org/eclipse/stardust/engine/core/cache/UserDomainsCache.java
@@ -135,7 +135,7 @@ public class UserDomainsCache extends AbstractCache<UserDomainBean>
       }
 
       UserDomainBean userDomain = UserDomainBean
-            .findByOID((short) oidKey.getOid());
+            .findByOID(oidKey.getOid());
       if (userDomain == null)
       {
          // remove phantom key
